Move crypto wrapper to storage directory
