+++ examples/main/src/packages/storage/LocalStorageMethod.js
@@ -1,4 +1,4 @@
-import { generateUUID } from "../../utilities";
+import { generateUUID } from "./cryptoWrapper";
 import { StorageMethod } from "./StorageMethod";
 
 /**
