+++ examples/main/src/packages/storage/index.js
@@ -2,5 +2,6 @@ import { StorageModelFactory } from "./StorageModel";
 import { StorageMethod } from "./StorageMethod";
 import { IndexedDBMethod } from "./IndexedDBMethod";
 import { LocalStorageMethod } from "./LocalStorageMethod";
+import { generateUUID } from "./cryptoWrapper";
 
-export { StorageModelFactory, StorageMethod, IndexedDBMethod, LocalStorageMethod };
+export { StorageModelFactory, StorageMethod, IndexedDBMethod, LocalStorageMethod, generateUUID };
