+++ examples/main/src/utilities/index.js
@@ -1,4 +1,1 @@
-import { generateUUID } from "./cryptoWrapper";
-
 export * from "./fieldValidators";
-export { generateUUID };
